+++ src/nexgen/command_line/nexus_generator.py
@@ -66,7 +66,7 @@ master_phil = freephil.parse(
 
     include scope nexgen.command_line.nxs_phil.goniometer_scope
 
-    include scope nexgen.command_line.nxs_phil.beamline_scope
+    include scope nexgen.command_line.nxs_phil.instrument_scope
 
     include scope nexgen.command_line.nxs_phil.detector_scope
 
@@ -99,7 +99,7 @@ demo_phil = freephil.parse(
 
     include scope nexgen.command_line.nxs_phil.goniometer_scope
 
-    include scope nexgen.command_line.nxs_phil.beamline_scope
+    include scope nexgen.command_line.nxs_phil.instrument_scope
 
     include scope nexgen.command_line.nxs_phil.detector_scope
 
@@ -135,7 +135,7 @@ meta_phil = freephil.parse(
 
     include scope nexgen.command_line.nxs_phil.goniometer_scope
 
-    include scope nexgen.command_line.nxs_phil.beamline_scope
+    include scope nexgen.command_line.nxs_phil.instrument_scope
 
     include scope nexgen.command_line.nxs_phil.detector_scope
 
@@ -226,7 +226,7 @@ def write_NXmx_cli(args):
     # Log information
     logger.info("Source information")
     logger.info(f"Facility: {source.name} - {source.type}.")
-    logger.info(f"Beamline: {source.beamline_name}")
+    logger.info(f"Beamline / Instrument: {source.beamline_name}")
 
     if timestamps[0] is not None:
         logger.info(f"Collection start time: {timestamps[0]}")
@@ -444,7 +444,7 @@ def write_demo_cli(args):
 
     logger.info("Source information")
     logger.info(f"Facility: {source.name} - {source.type}.")
-    logger.info(f"Beamline: {source.beamline_name}")
+    logger.info(f"Beamline / Instrument: {source.beamline_name}")
 
     logger.info("Coordinate system: %s" % cf)
     if cf == "imgcif":
@@ -729,7 +729,7 @@ def write_with_meta_cli(args):
     # Log information
     logger.info("Source information")
     logger.info(f"Facility: {source.name} - {source.type}.")
-    logger.info(f"Beamline: {source.beamline_name}")
+    logger.info(f"Beamline / Instrument: {source.beamline_name}")
 
     if timestamps[0] is not None:
         logger.info(f"Collection start time: {timestamps[0]}")