+++ src/nexgen/nxs_write/EDNexusWriter.py
@@ -141,7 +141,7 @@ def ED_call_writers(
         nxsfile,
         beam,
         attenuator,
-        source["beamline_name"],
+        source,
     )
 
     # NXdetector: /entry/instrument/detector